requirements.txt
scikit-learn==0.24.0
torch_scatter==2.0.5
torch_sparse==0.6.8
torch_cluster==1.5.8
torch_geometric==1.6.3
tqdm==4.55.1